refactor: renamed all instances of getProducts dir to products dir to improve readability
